+++ scriptconfig/config.py
@@ -282,6 +282,9 @@ class Config(ub.NiceRepr, DictLike, metaclass=MetaConfig):
         epilog (str): A class attribute that if specified will add an epilog
             section to the help text.
 
+    SeeAlso:
+        :class:`scriptconfig.DataConfig`
+
     Example:
         >>> # Inherit from `Config` and assign `__default__`
         >>> import scriptconfig as scfg
@@ -852,6 +855,15 @@ class Config(ub.NiceRepr, DictLike, metaclass=MetaConfig):
             >>> parser.print_help()
             >>> x = parser.parse_known_args()
 
+        Example:
+            >>> import scriptconfig as scfg
+            >>> import pytest
+            >>> class EmptyConfig(scfg.Config):
+            >>>     ...
+            >>> self = EmptyConfig()
+            >>> with pytest.raises(Exception) as ex:
+            >>>     self._read_argv(argv=32132)
+
         Ignore:
             >>> # Weird cases
             >>> self = MyConfig()
@@ -890,10 +902,31 @@ class Config(ub.NiceRepr, DictLike, metaclass=MetaConfig):
             else:
                 argcomplete_mod.autocomplete(parser)
 
-        if strict:
-            ns = parser.parse_args(argv).__dict__
-        else:
-            ns = parser.parse_known_args(argv)[0].__dict__
+        try:
+            if strict:
+                ns = parser.parse_args(argv).__dict__
+            else:
+                ns = parser.parse_known_args(argv)[0].__dict__
+        except (ValueError, TypeError) as ex:
+            # For errors (like ValueError) where its probably a programmer
+            # error and not a user error, give the debugger some information
+            # about the scriptconfig object.
+            from scriptconfig.util import util_exception
+            # TODO: figure out argv that triggers a value error so we can add a test
+            note = ub.codeblock(
+                f'''
+                Error while attempting to parse arguments in _read_argv
+
+                Context:
+                    argv = {argv!r}
+                    special_options = {special_options!r}
+                    strict = {strict!r}
+                    autocomplete = {autocomplete!r}
+                    self = {self!r}
+                ''')
+            print(note)
+            ex = util_exception.add_exception_note(ex, note)
+            raise ex
 
         special_ns_keys = ['config', 'dump', 'dumps']
         if special_options:
@@ -1603,7 +1636,7 @@ class Config(ub.NiceRepr, DictLike, metaclass=MetaConfig):
         if parser is None:
             parserkw = self._parserkw()
             # parser = argparse.ArgumentParser(**parserkw)
-            parser = argparse_ext.CompatArgumentParser(**parserkw)
+            parser = argparse_ext.ExtendedArgumentParser(**parserkw)
 
         # Use custom action used to mark which values were explicitly set on
         # the commandline